+++ StockAnalysis/source/SystemMessage.h
@@ -15,7 +15,7 @@ public:
 
   void PushMessage(CString str);
   CString PopMessage(void);
-  long GetDequeSize(void);
+  size_t GetDequeSize(void);
 
   void Display(COutputList* pOutputList, CString strTime);
 
@@ -34,37 +34,37 @@ public:
 
   void PushInformationMessage(CString str) { m_SystemInformation.PushMessage(str); }
   CString PopInformationMessage(void) { return m_SystemInformation.PopMessage(); }
-  long GetInformationDequeSize(void) { return m_SystemInformation.GetDequeSize(); }
+  size_t GetInformationDequeSize(void) { return m_SystemInformation.GetDequeSize(); }
   void DisplayInformation(COutputList* pOutputList, CString strTime) { m_SystemInformation.Display(pOutputList, strTime); }
 
   void PushDayLineInfoMessage(CString str) { m_DayLineInformation.PushMessage(str); }
   CString PopDayLineInfoMessage(void) { return m_DayLineInformation.PopMessage(); }
-  long GetDayLineInfoDequeSize(void) { return m_DayLineInformation.GetDequeSize(); }
+  size_t GetDayLineInfoDequeSize(void) { return m_DayLineInformation.GetDequeSize(); }
   void DisplayDayLineInfo(COutputList* pOutputList, CString strTime) { m_DayLineInformation.Display(pOutputList, strTime); }
 
   void PushTransactionMessage(CString str) { m_Transaction.PushMessage(str); }
   CString PopTransactionMessage(void) { return m_Transaction.PopMessage(); }
-  long GetTransactionDequeSize(void) { return m_Transaction.GetDequeSize(); }
+  size_t GetTransactionDequeSize(void) { return m_Transaction.GetDequeSize(); }
   void DisplayTransaction(COutputList* pOutputList, CString strTime) { m_Transaction.Display(pOutputList, strTime); }
 
   void PushCancelSellMessage(CString str) { m_CancelSell.PushMessage(str); }
   CString PopCancelSellMessage(void) { return m_CancelSell.PopMessage(); }
-  long GetCancelSellDequeSize(void) { return m_CancelSell.GetDequeSize(); }
+  size_t GetCancelSellDequeSize(void) { return m_CancelSell.GetDequeSize(); }
   void DisplayCancelSell(COutputList* pOutputList, CString strTime) { m_CancelSell.Display(pOutputList, strTime); }
 
   void PushCancelBuyMessage(CString str) { m_CancelBuy.PushMessage(str); }
   CString PopCancelBuyMessage(void) { return m_CancelBuy.PopMessage(); }
-  long GetCancelBuyDequeSize(void) { return m_CancelBuy.GetDequeSize(); }
+  size_t GetCancelBuyDequeSize(void) { return m_CancelBuy.GetDequeSize(); }
   void DisplayCancelBuy(COutputList* pOutputList, CString strTime) { m_CancelBuy.Display(pOutputList, strTime); }
 
   void PushTrace2Message(CString str) { m_Trace2.PushMessage(str); }
   CString PopTrace2Message(void) { return m_Trace2.PopMessage(); }
-  long GetTrace2DequeSize(void) { return m_Trace2.GetDequeSize(); }
+  size_t GetTrace2DequeSize(void) { return m_Trace2.GetDequeSize(); }
   void DisplayTrace2(COutputList* pOutputList, CString strTime) { m_Trace2.Display(pOutputList, strTime); }
 
   void PushInnerSystemInformationMessage(CString str) { m_InnerSystemInformation.PushMessage(str); }
   CString PopInnerSystemInformationMessage(void) { return m_InnerSystemInformation.PopMessage(); }
-  long GetInnerSystemInformationDequeSize(void) { return m_InnerSystemInformation.GetDequeSize(); }
+  size_t GetInnerSystemInformationDequeSize(void) { return m_InnerSystemInformation.GetDequeSize(); }
   void DisplayInnerSystemInformation(COutputList* pOutputList, CString strTime) { m_InnerSystemInformation.Display(pOutputList, strTime); }
 
 protected:
